+++ R/tbl-df.r
@@ -35,19 +35,6 @@ tbl_df <- function(data) {
   as_data_frame(data)
 }
 
-#' @export
-as.tbl.data.frame <- function(x, ...) {
-  tbl_df(x)
-}
-
-#' @export
-tbl_vars.data.frame <- function(x) names(x)
-
-#' @export
-same_src.data.frame <- function(x, y) {
-  is.data.frame(y)
-}
-
 # Standard data frame methods --------------------------------------------------
 
 #' @export